pyproject.toml
[tool.poetry]
name = "gt-bot"
version = "0.1.0"
description = "gt-bot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-beta.1"
nonebot-adapter-onebot = "^2.0.0-beta.1"
nonebot-plugin-setu = "^1.1.1"
pydantic = "^1.9.0"
Pillow = "^9.1.1"
saucenao-api = "^2.4.0"
pytz = "^2022.1"
beautifulsoup4 = "^4.11.1"
pandas = "^1.4.2"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = ["nonebot_plugin_setu"]
plugin_dirs = ["gt/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
